Move mario animations to player_animations file
